[ARV-146] refactor: 차량 대절 API swagger ui 개선
